requirements.txt
django==5.0.0
line-bot-sdk==3.17.1
python-dotenv==1.1.0
requests==2.32.3
openai==1.77.0 